Fix: capture_matches handle exists: / correctly
Previsouly capture_matches returned and Option<>.
This need to be change to a result as this function could fail
when exists: mechanisms contain a `/` character as they are not allowed.

Related code needed to be updated to ensure that code the code still
works as expected.

Additional tests will need to be added and updated moving forward to
ensure no other issues have been introduced.

feat: add InvalidMechanimsFormatByKind Error

This error reports the error message but also includes the mechanims
KIND. So that it is clear which mechanism created the error.

Related error handlding code was also updated.

<insert any additional information, such as references or issue numbers>

refactor tests: regex_capture `a` mechanisms

refactor tests: capture_matches for exists

refactor tests for capture_matches for mx

Refactor tests ptr

Add tests for ptr that should fail