feat(iterators) Implement iterators for Spf
Add some support for iterating over the Spf Struct.

This might change as it can not support the verion field. This means the version
field needs to be accessed before the iterator is called.

It might be better to returin the Generic T instead of Mechanism<T>

test(iter) Test for iter for Spf

Add test to confirm that the Spf iter works as expected.

doc(iterator)