refactor: Rewrite Display trait for Spf
rewrite Display trait for Spf
Previously this would return the value stored in `source`
However if the Spf is created using `SpfBuilder` then source is empty.

The new implemention now check is `source` is empty. If is not emtpy then `source` is used.
If `source` is empty, the string is built using `version` and by calling `iter()` on self.